tests: refactor kotlin tests to not depend on other tests
